Fix issues in the code review - Set the default values of Origin/Cache_SelfTestInterval - Move self-monitoring.go to xrootd pkg, avoid creating a new pkg - Use filepath.Join to construct the path - Other minor improvements
