+++ xrootd/launch.go
@@ -39,6 +39,7 @@ import (
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/daemon"
 	"github.com/pelicanplatform/pelican/param"
+	"github.com/pelicanplatform/pelican/server_utils"
 )
 
 type (
@@ -159,9 +160,9 @@ func makeUnprivilegedXrootdLauncher(daemonName string, configPath string, isCach
 		result.ExtraEnv = append(result.ExtraEnv, "XRDHTTP_PELICAN_CERT_FILE="+filepath.Join(xrootdRun, "copied-tls-creds.crt"))
 		result.ExtraEnv = append(result.ExtraEnv, "XRDHTTP_PELICAN_INFO_FD="+strconv.Itoa(result.fds[1]))
 
-		basePath := param.Cache_NamespaceLocation.GetString() + "/pelican/monitoring/selfTest"
-		testFileLocation := basePath + "/self-test-cache-server.txt"
-		testFileCinfoLocation := basePath + "/self-test-cache-server.txt.cinfo"
+		basePath := filepath.Join(param.Cache_NamespaceLocation.GetString(), server_utils.MonitoringBaseNs, "selfTest")
+		testFileLocation := filepath.Join(basePath, "self-test-cache-server.txt")
+		testFileCinfoLocation := filepath.Join(basePath, "self-test-cache-server.txt.cinfo")
 		result.ExtraEnv = append(result.ExtraEnv, "XRDHTTP_PELICAN_CACHE_SELF_TEST_FILE="+testFileLocation)
 		result.ExtraEnv = append(result.ExtraEnv, "XRDHTTP_PELICAN_CACHE_SELF_TEST_FILE_CINFO="+testFileCinfoLocation)
 	}
